feat: improve report modal

+++ app/src/main/java/com/insa/iss/safecityforcyclists/database/LocalReport.kt
@@ -9,7 +9,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "local_reports")
 data class LocalReport(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    @ColumnInfo(name = "timestamp") val timestamp: Int,
+    @ColumnInfo(name = "timestamp") val timestamp: Long,
     @ColumnInfo(name = "distance") val distance: Double,
     @ColumnInfo(name = "object_speed") val objectSpeed: Double,
     @ColumnInfo(name = "bicycle_speed") val bicycleSpeed: Double,
